+++ bulkimport/__init__.py
@@ -1,0 +1,5 @@
+try:
+    import importlib
+    importlib.import_module('parsers')
+except ImportError:
+    pass