+++ geotrek/common/forms.py
@@ -1,5 +1,7 @@
+# -*- coding: utf-8 -*-
 from django import forms as django_forms
 from django.db.models.fields.related import ForeignKey, ManyToManyField, FieldDoesNotExist
+from django.utils.translation import ugettext_lazy as _
 
 import floppyforms as forms
 from mapentity.forms import MapEntityForm
@@ -8,6 +10,10 @@ from geotrek.authent.models import (default_structure, StructureRelated,
                                     StructureRelatedQuerySet)
 
 from .mixins import NoDeleteMixin
+
+from crispy_forms.helper import FormHelper
+from crispy_forms.layout import Layout, Div, Submit
+from crispy_forms.bootstrap import FormActions
 
 
 class CommonForm(MapEntityForm):
@@ -68,3 +74,39 @@ class CommonForm(MapEntityForm):
 
         for name, field in self.fields.items():
             self.filter_related_field(name, field)
+
+
+class ImportDatasetForm(django_forms.Form):
+    parser = forms.TypedChoiceField(
+        label=_('Parser'),
+        widget=forms.RadioSelect,
+        required=True,
+    )
+    zipfile = forms.FileField(
+        label=_('File'),
+        required=True,
+        widget=forms.FileInput
+    )
+
+    def __init__(self, choices=None, *args, **kwargs):
+            super(ImportDatasetForm, self).__init__(*args, **kwargs)
+
+            self.fields['parser'].choices = choices
+
+            self.helper = FormHelper()
+            self.helper.layout = Layout(
+                Div(
+                    Div(
+                        'parser',
+                        'zipfile',
+                    ),
+                    FormActions(
+                        Submit('submit', _("Upload"), css_class='button white')
+                    ),
+                    css_class='file-attachment-form',
+                )
+            )
+
+    def clean_zipfile(self):
+        if self.cleaned_data['zipfile'].content_type != "application/zip":
+            raise django_forms.ValidationError(_("File must be of ZIP type."), code='invalid')